+++ tests/import_snapshot_tests.rs
@@ -5,14 +5,14 @@ pub mod common;
 
 use anyhow::Result;
 
-use crate::common::{cli, create_tmp_config, CommonEnv};
+use crate::common::{create_tmp_config, daemon, CommonEnv};
 
 #[test]
 fn importing_bad_snapshot_should_fail() -> Result<()> {
     let (config_file, data_dir) = create_tmp_config()?;
     let temp_file = data_dir.path().join("bad-snapshot.car");
     std::fs::write(&temp_file, "bad-snapshot")?;
-    cli()?
+    daemon()?
         .common_env()
         .arg("--rpc-address")
         .arg("127.0.0.1:0")